+++ src/snowflake/snowpark/_internal/ast/utils.py
@@ -340,7 +340,12 @@ def build_proto_from_struct_type(
     expr.structured = schema.structured
     for field in schema.fields:
         ast_field = expr.fields.list.add()
-        field.column_identifier._fill_ast(ast_field.column_identifier)  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "ColumnIdentifier" has no attribute "_fill_ast"
+        if isinstance(field.original_column_identifier, str):
+            ast_field.column_identifier.sp_column_name.name = (
+                field.original_column_identifier
+            )
+        else:
+            field.column_identifier._fill_ast(ast_field.column_identifier.sp_column_identifier)  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "ColumnIdentifier" has no attribute "_fill_ast"
         field.datatype._fill_ast(ast_field.data_type)  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "DataType" has no attribute "_fill_ast"
         ast_field.nullable = field.nullable
 
@@ -888,7 +893,7 @@ def build_expr_from_snowpark_column_or_table_fn(
 
 # TODO(SNOW-1491199) - This method is not covered by tests until the end of phase 0. Drop the pragma when it is covered.
 def fill_ast_for_column(  # type: ignore[no-untyped-def] # TODO(SNOW-1491199) # Function is missing a type annotation for one or more arguments
-    expr: proto.Expr, name1: str, name2: Optional[str], fn_name="col"
+    expr: proto.Expr, name1: str, name2: Optional[str], fn_name=None
 ) -> None:  # pragma: no cover
     """
     Fill in expr node to encode Snowpark Column created through col(...) / column(...).
@@ -899,19 +904,29 @@ def fill_ast_for_column(  # type: ignore[no-untyped-def] # TODO(SNOW-1491199) # 
         fn_name: alias to use when encoding Snowpark column (should be "col" or "column").
 
     """
+    # Internal calls to the Column constructor will set the fn_name parameter only if an AST for a builtin_fn needs to be generated (e.g. calls from functions.col or functions.column)
+    # Otherwise, all internal calls to the Column constructor will leave the fn_name parameter as None (e.g. calls from Column._to_col_if_<other> private methods), meaning that only
+    # a string or other type was provided, and we internally generate Snowpark.Column instance with the name provided, which should not generate a builtin_fn AST.
+    if fn_name is None:
+        # Handle the special case * (as a SQL column expr).
+        if name2 == "*":
+            ast = with_src_position(expr.sp_column_sql_expr)  # type: ignore[arg-type] # TODO(SNOW-1491199) # Argument 1 to "with_src_position" has incompatible type "SpColumnSqlExpr"; expected "Expr"
+            ast.sql = "*"  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "Expr" has no attribute "sql"
+            if name1 is not None:
+                ast.df_alias.value = name1  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "Expr" has no attribute "df_alias"
+            return expr  # type: ignore[return-value] # TODO(SNOW-1491199) # No return value expected
 
-    # Handle the special case * (as a SQL column expr).
-    if name2 == "*":
-        ast = with_src_position(expr.sp_column_sql_expr)  # type: ignore[arg-type] # TODO(SNOW-1491199) # Argument 1 to "with_src_position" has incompatible type "SpColumnSqlExpr"; expected "Expr"
-        ast.sql = "*"  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "Expr" has no attribute "sql"
-        if name1 is not None:
-            ast.df_alias.value = name1  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "Expr" has no attribute "df_alias"
-        return expr  # type: ignore[return-value] # TODO(SNOW-1491199) # No return value expected
+        if name1 == "*" and name2 is None:
+            ast = with_src_position(expr.sp_column_sql_expr)  # type: ignore[arg-type] # TODO(SNOW-1491199) # Argument 1 to "with_src_position" has incompatible type "SpColumnSqlExpr"; expected "Expr"
+            ast.sql = "*"  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "Expr" has no attribute "sql"
+            return expr  # type: ignore[return-value] # TODO(SNOW-1491199) # No return value expected
 
-    if name1 == "*" and name2 is None:
-        ast = with_src_position(expr.sp_column_sql_expr)  # type: ignore[arg-type] # TODO(SNOW-1491199) # Argument 1 to "with_src_position" has incompatible type "SpColumnSqlExpr"; expected "Expr"
-        ast.sql = "*"  # type: ignore[attr-defined] # TODO(SNOW-1491199) # "Expr" has no attribute "sql"
-        return expr  # type: ignore[return-value] # TODO(SNOW-1491199) # No return value expected
+        # Correctly captures all cases in which a Snowpark public API accepts a string or a Column object. If a Column object was provided, then its AST should have been used,
+        # otherwise the string value will be passed to the Column constructor internally, but the fn_name parameter will be None, meaning only a string was provided to the public API.
+        if name2 is None:
+            ast = with_src_position(expr.string_val)
+            ast.v = name1
+            return expr
 
     # Regular form (without *): build as function ApplyExpr.
     kwargs = (
@@ -930,7 +945,7 @@ def fill_ast_for_column(  # type: ignore[no-untyped-def] # TODO(SNOW-1491199) # 
 
 # TODO(SNOW-1491199) - This method is not covered by tests until the end of phase 0. Drop the pragma when it is covered.
 def create_ast_for_column(  # type: ignore[no-untyped-def] # TODO(SNOW-1491199) # Function is missing a type annotation for one or more arguments
-    name1: str, name2: Optional[str], fn_name="col"
+    name1: str, name2: Optional[str], fn_name=None
 ) -> proto.Expr:  # pragma: no cover
     """
     Helper function to create Ast for Snowpark Column. Cf. fill_ast_for_column on parameter details.