+++ src/snowflake/snowpark/_internal/server_connection.py
@@ -49,7 +49,10 @@ from snowflake.snowpark._internal.analyzer.snowflake_plan import (
 )
 from snowflake.snowpark._internal.ast.utils import DATAFRAME_AST_PARAMETER
 from snowflake.snowpark._internal.error_message import SnowparkClientExceptionMessages
-from snowflake.snowpark._internal.telemetry import TelemetryClient
+from snowflake.snowpark._internal.telemetry import (
+    TelemetryClient,
+    get_plan_telemetry_metrics,
+)
 from snowflake.snowpark._internal.utils import (
     create_rlock,
     create_thread_local,
@@ -776,6 +779,12 @@ class ServerConnection:
                         case_sensitive=case_sensitive,
                         **kwargs,
                     )
+
+            if plan.session._collect_snowflake_plan_telemetry_at_critical_path:
+                self._telemetry_client.send_plan_metrics_telemetry(
+                    session_id=self.get_session_id(),
+                    data=get_plan_telemetry_metrics(plan),
+                )
 
         if result is None:
             raise SnowparkClientExceptionMessages.SQL_LAST_QUERY_RETURN_RESULTSET()